made tablebutton into an actual class again
