+++ client/components/FloorPlan.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Grid from '@material-ui/core/Grid';
 import Box from '@material-ui/core/Box';
-import { TableButton } from './TableButton';
+import TableButton from './TableButton';
 
 export default class FloorPlan extends React.Component {
   constructor(props) {
@@ -22,57 +22,13 @@ export default class FloorPlan extends React.Component {
       });
   }
 
-  parseTimeSeated(timeSeated) {
-    if (timeSeated) {
-      const splitTime = timeSeated.split('T');
-      const furtherSplit = splitTime[1].split(':');
-      let [hours, minutes] = furtherSplit;
-      hours = parseInt(hours, 10);
-      // this is for timezone.  Idk why 5 works
-      hours += 5;
-      if (hours > 12) {
-        hours -= 12;
-      }
-      let amPM = 'AM';
-      if (hours > 11) {
-        amPM = 'PM';
-      }
-      if (hours === 24 || hours === 0) {
-        hours = 12;
-        amPM = 'AM';
-      } else if (hours > 12) {
-        hours -= 12;
-      }
-      return `${hours}:${minutes} ${amPM}`;
-    }
-    return '';
-  }
-
   render() {
     const { tables } = this.state;
     const tableList = tables.map(table => {
-      const { tableId, tableStatus, timeSeated } = table;
-      let color;
-      switch (tableStatus) {
-        case (0):
-          color = 'default';
-          break;
-        case (1):
-          color = 'primary';
-          break;
-        case (2):
-          color = 'secondary';
-          break;
-        default:
-      }
-      const children = <div>{`T${tableId}`}<div>{this.parseTimeSeated(timeSeated)}</div></div>;
       return (
         <TableButton
-          key={tableId}
+          key={table.tableId}
           tableData={table}
-          color={color}
-          text={children}
-          status={table}
           viewDialog={this.props.viewDialog}
           dialogOpen={this.props.dialogOpen}
         />
